+++ src/features/account/model/useJoin.ts
@@ -4,19 +4,11 @@ import { useForm } from 'react-hook-form';
 import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useRouter } from 'next/navigation';
-import { useState } from 'react';
 
 export const useJoin = () => {
-  const [isEmailDuplicated, setIsEmailDuplicated] = useState(true);
-  const [otpVisible, setOtpVisible] = useState(false);
-  const [verificationComplete, setVerificationComplete] = useState(false);
-  const [otpInput, setOtpInput] = useState('');
-  const [otpError, setOtpError] = useState(false);
-
   const userNameRegex = /^[가-힣]+$/;
   const passwordRegex =
     /^(?=.*[a-z])(?=.*\d)(?=.*[@$!%*?&])[a-z\d@$!%*?&]{8,20}$/;
-  const emailRegex = /^[a-zA-Z0-9._%+-]+@mju\.ac\.kr$/;
 
   const userSchema = z.object({
     userName: z
@@ -24,10 +16,6 @@ export const useJoin = () => {
       .min(2, { message: '2글자 이상 입력해주세요.' })
       .max(15, { message: '15글자 이하로 입력해 주세요.' })
       .regex(userNameRegex, { message: '한글 조합만 사용 가능합니다.' }),
-    email: z
-      .string()
-      .email({ message: '이메일 형식으로 입력해 주세요.' })
-      .regex(emailRegex, { message: '@mju.ac.kr 형식이어야 합니다.' }),
     password: z
       .string()
       .min(8, { message: '8글자 이상 입력해주세요.' })
@@ -43,7 +31,6 @@ export const useJoin = () => {
     resolver: zodResolver(userSchema),
     defaultValues: {
       userName: '',
-      email: '',
       password: '',
       college: '',
       department: '',
@@ -54,39 +41,11 @@ export const useJoin = () => {
     formState: { errors },
   } = form;
 
-  const isEmailValid = !errors.email && form.getValues('email').length > 0;
-
-  const emailDuplicationCheck = () => {
-    const email = form.getValues('email');
-    if (email) {
-      if (email.includes('test')) {
-        setIsEmailDuplicated(true);
-      } else {
-        setIsEmailDuplicated(false);
-      }
-    }
-  };
-
-  const isVerificationButtonEnabled = !isEmailDuplicated && isEmailValid;
-
-  const handleVerifyOtp = () => {
-    //임시로 111111과 일치할 경우로 달아놓음. api 자리
-    if (otpInput === '111111') {
-      setVerificationComplete(true);
-      setOtpError(false);
-    } else {
-      setVerificationComplete(false);
-      setOtpError(true);
-    }
-  };
-
   const isSubmitButtonEnabled =
     !errors.userName &&
     form.getValues('college') &&
     form.getValues('department') &&
-    !errors.password &&
-    !isEmailDuplicated &&
-    verificationComplete;
+    !errors.password;
 
   const router = useRouter();
 
@@ -97,18 +56,7 @@ export const useJoin = () => {
 
   return {
     form,
-    isEmailValid,
-    emailDuplicationCheck,
-    isEmailDuplicated,
-    verificationComplete,
-    otpVisible,
-    handleVerifyOtp,
-    setOtpVisible,
-    setOtpInput,
-    otpInput,
-    otpError,
     onSubmit,
     isSubmitButtonEnabled,
-    isVerificationButtonEnabled,
   };
 };
